# BikeRental/BikeRental/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "BikeRental by Michael Hitzker",
        "version": "v1"
    },
    "paths": {
...
        "/api/Rentals/start": {
            "post": {
                "tags": [
                    "Rentals"
                ],
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "bikeId",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/Rentals/end": {
            "post": {
                "tags": [
                    "Rentals"
                ],
                "parameters": [
                    {
                        "name": "rentalId",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/Rentals/pay": {
            "post": {
                "tags": [
                    "Rentals"
                ],
                "parameters": [
                    {
                        "name": "rentalId",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
...
}